Move waypoints code into a separate file.
