+++ Commands/building/CmdClick.cs
@@ -61,7 +61,7 @@ namespace MCGalaxy.Commands
             }
 
             p.lastCMD = "click";
-            p.manualChange(click[0], click[1], click[2], 0, Block.rock);
+            p.ManualChange(click[0], click[1], click[2], 0, Block.rock);
             Player.SendMessage(p, "Clicked &b(" + click[0] + ", " + click[1] + ", " + click[2] + ")");
         }
         public override void Help(Player p)
