+++ MCGalaxy_.csproj
@@ -444,6 +444,7 @@
     <Compile Include="Levels\Physics\ZombiePhysics.cs" />
     <Compile Include="Player\Chat.cs" />
     <Compile Include="Player\Player.Timers.cs" />
+    <Compile Include="Player\Waypoint.cs" />
     <Compile Include="Plugins\Enums.cs" />
     <Compile Include="Plugins\Events\GroupEvents.cs" />
     <Compile Include="Plugins\Events\LevelEvents.cs" />
